refactor(timeline): extract fade-edit state machines into useFadeEditors
Pull the four parallel fade-edit state machines (video fade, audio fade,
audio fade curve, audio volume) plus their refs, commit effects, and
~500 lines of pointer handlers out of TimelineItem into a single hook.

The audio volume CSS-variable preview stays imperative (no React state
on the drag hot path), and all four state machines preserve their
isCommitting cycle that auto-clears once persisted values match the
preview.

index.tsx shrinks from 3,588 to ~2,800 lines.
